chore: move AppConstants to utils package
